[bug] index.html was unable to retrieve js bundle.
We got the error message:
Failed to load module script: Expected a JavaScript module script but the server responded with a MIME type of "text/plain". Strict MIME type checking is enforced for module scripts per HTML spec.
after startup of our container and going to localhost to access our website. the index.html was succesffuly being received if the error message was trying to get js files. we resolve the problem by addig typing for the nginx config so that the js modules are served by nginx w the right headers/typing.